+++ src/components/App/App.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { shallow, mount } from 'enzyme';
-import App from './App';
+import App from './App.jsk';
 
 
 describe('App', () => {
