+++ src/components/Container/Container.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Card from '../Card/Card';
+import Card from '../Card/Card.jsx';
 import './Container.css';
 
 const Container = ({ data, favoriteCard }) => {
