--- src/components/Header/Header.css
@@ -1,28 +1,0 @@
-header {
-  display: flex;
-  justify-content: space-around;
-  align-items: center
-}
-
-.change-btn {
-  display: flex;
-  justify-content: space-around;
-  margin-bottom: 50px;
-}
-
-.btns {
-  border-radius: 15px;
-  cursor: pointer;
-  font-family: 'Raleway';
-  font-size: 16px;
-  opacity: 0.8;
-  padding: 10px;
-  width: 150px;
-  margin: 1%;
-}
-
-.btns:hover {
-  background-color: #8DB4ED;
-  font-weight: bold;
-  outline: none;
-}
